Realized LikeSystem cerberus-hippogriff12/leto007/BJS2-96275

+++ src/main/java/faang/school/postservice/dto/post/PostV2Dto.java
@@ -25,6 +25,6 @@ public record PostV2Dto(Long id,
                       LocalDateTime createdAt,
                         @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
                       LocalDateTime updatedAt,
-                        Integer likesCount
+                        Long likesCount
 ) {
 }
